+++ spec/support/configs/aruba.rb
@@ -1,4 +1,4 @@
-require 'aruba/cucumber'
+require 'aruba/rspec'
 
 Aruba.configure do |config|
   config.exit_timeout                          = 120